Refactored the code to enable the user to update the product rating

+++ database/migrations/2020_05_10_151731_create_product_rating_table.php
@@ -14,6 +14,7 @@ class CreateProductRatingTable extends Migration
     public function up()
     {
         Schema::create('product_rating', function (Blueprint $table) {
+
             $table->primary(['product_id', 'rating_id', 'user_id']);
 
             $table->foreignId('product_id')->constrained()
@@ -22,8 +23,7 @@ class CreateProductRatingTable extends Migration
             $table->foreignId('rating_id')->constrained()
                 ->onDelete('cascade');
 
-            $table->unsignedBigInteger('user_id');
-
+            $table->unsignedInteger('user_id');
         });
     }
